Refactor WrappedArgumentsNode => ParameterizedNode::WrappedArguments
