+++ lib/rubocop/ast.rb
@@ -18,7 +18,6 @@ require_relative 'ast/node/mixin/numeric_node'
 require_relative 'ast/node/mixin/parameterized_node'
 require_relative 'ast/node/mixin/predicate_operator_node'
 require_relative 'ast/node/mixin/basic_literal_node'
-require_relative 'ast/node/mixin/wrapped_arguments_node'
 require_relative 'ast/node/alias_node'
 require_relative 'ast/node/and_node'
 require_relative 'ast/node/args_node'
